chore: front and back compartilhando mesmo schema
