+++ packages/server/package.json
@@ -8,6 +8,7 @@
     "db:migrate": "npx drizzle-kit migrate"
   },
   "dependencies": {
+    "@repo/zod-types": "*",
     "@hono/zod-validator": "^0.2.2",
     "@libsql/client": "^0.10.0",
     "bcryptjs": "^2.4.3",
