+++ packages/ui/package.json
@@ -15,6 +15,7 @@
   "devDependencies": {
     "@chromatic-com/storybook": "^1.8.0",
     "@repo/eslint-config": "*",
+    "@repo/zod-types": "*",
     "@repo/typescript-config": "*",
     "@storybook/addon-essentials": "^8.2.9",
     "@storybook/addon-interactions": "^8.2.9",
